# backend3/docs/swagger.yaml
basePath: /api
definitions:
...
  controller_k8s.CreateClusterRequest:
    properties:
      clusterType:
        type: string
      description:
        type: string
      endpoint:
        type: string
      kubeconfig:
        type: string
      name:
        type: string
      nodeCount:
        type: integer
      region:
        type: string
    required:
    - endpoint
    - kubeconfig
    - name
    type: object
...
  controller_k8s.DeleteClusterRequest:
    properties:
      confirmName:
        type: string
    required:
    - confirmName
    type: object
...
  controller_k8s.UpdateClusterRequest:
    properties:
      clusterType:
        type: string
      description:
        type: string
      endpoint:
        type: string
      kubeconfig:
        type: string
      name:
        type: string
      nodeCount:
        type: integer
      region:
        type: string
      status:
        type: integer
    type: object
...
  oneops_backend3_pkg_utils.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:8082
info:
  contact: {}
  description: |-
    OneOps 一体化运维平台后端 API
    包含系统管理、CMDB、K8s、审计、监控、应用授权等模块
  title: OneOps API
  version: "1.0"
paths:
...
  /k8s/clusters:
    get:
      description: 分页获取当前用户有权限的 K8s 集群列表，支持按名称、状态、类型筛选
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      - description: 集群名称
        in: query
        name: name
        type: string
      - description: 状态（0/1）
        in: query
        name: status
        type: string
      - description: 集群类型
        in: query
        name: clusterType
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取集群列表失败
          schema:
            $ref: '#/definitions/oneops_backend3_pkg_utils.Response'
      security:
      - BearerAuth: []
      summary: 获取集群列表
      tags:
      - K8s-集群管理
    post:
      consumes:
      - application/json
      description: 创建一个新的 K8s 集群配置（含 kubeconfig）
      parameters:
      - description: 创建集群请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller_k8s.CreateClusterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建集群失败
          schema:
            $ref: '#/definitions/oneops_backend3_pkg_utils.Response'
      security:
      - BearerAuth: []
      summary: 创建集群
      tags:
      - K8s-集群管理
  /k8s/clusters/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定 K8s 集群（需要二次确认，传入集群名称）
      parameters:
      - description: 集群 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 删除集群请求（需 confirmName 二次确认）
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller_k8s.DeleteClusterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 删除集群失败
          schema:
            $ref: '#/definitions/oneops_backend3_pkg_utils.Response'
      security:
      - BearerAuth: []
      summary: 删除集群
      tags:
      - K8s-集群管理
    get:
      description: 根据 ID 获取指定 K8s 集群的详细信息
      parameters:
      - description: 集群 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取集群详情失败
          schema:
            $ref: '#/definitions/oneops_backend3_pkg_utils.Response'
      security:
      - BearerAuth: []
      summary: 获取集群详情
      tags:
      - K8s-集群管理
    put:
      consumes:
      - application/json
      description: 更新指定 K8s 集群的配置信息
      parameters:
      - description: 集群 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新集群请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller_k8s.UpdateClusterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新集群失败
          schema:
            $ref: '#/definitions/oneops_backend3_pkg_utils.Response'
      security:
      - BearerAuth: []
      summary: 更新集群
      tags:
      - K8s-集群管理
...
  /k8s/clusters/{id}/namespaces:
    get:
      description: 获取指定 K8s 集群的命名空间列表
      parameters:
      - description: 集群 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取命名空间列表失败
          schema:
            $ref: '#/definitions/oneops_backend3_pkg_utils.Response'
      security:
      - BearerAuth: []
      summary: 获取集群命名空间列表
      tags:
      - K8s-集群管理
  /k8s/clusters/{id}/nodes:
    get:
      description: 获取指定 K8s 集群的节点列表
      parameters:
      - description: 集群 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取节点列表失败
          schema:
            $ref: '#/definitions/oneops_backend3_pkg_utils.Response'
      security:
      - BearerAuth: []
      summary: 获取集群节点列表
      tags:
      - K8s-集群管理
...
  /k8s/clusters/{id}/test:
    post:
      description: 测试指定 K8s 集群的连通性
      parameters:
      - description: 集群 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 连接测试失败
          schema:
            $ref: '#/definitions/oneops_backend3_pkg_utils.Response'
      security:
      - BearerAuth: []
      summary: 测试集群连接
      tags:
      - K8s-集群管理
  /k8s/clusters/{id}/users:
    get:
      description: 获取指定 K8s 集群下的用户列表及权限信息
      parameters:
      - description: 集群 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取集群用户列表失败
          schema:
            $ref: '#/definitions/oneops_backend3_pkg_utils.Response'
      security:
      - BearerAuth: []
      summary: 获取集群用户列表
      tags:
      - K8s-集群管理
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 输入 Bearer {token}，token 由 /api/login 获取
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"